Move Physical properties to schema
