+++ transmission/transmission_plugin/uv_vis_nir_transmission_plugin/src/transmission/schema.py
@@ -57,11 +57,10 @@ from nomad.units import ureg
 from nomad.datamodel.metainfo.plot import (
     PlotSection,
     PlotlyFigure,
-)
+from nomad_material_processing import Parallelepiped
 
 from transmission.readers import read_asc
-from transmission.utils import merge_sections, create_archive
-from transmission.material_systems import PhysicalProperties
+from transmission.utils import create_archive, merge_sections
 
 if TYPE_CHECKING:
     from nomad.datamodel.datamodel import EntryArchive
@@ -69,6 +68,51 @@ if TYPE_CHECKING:
 
 
 m_package = Package(name='nomad_transmission')
+
+
+class TransmissionParallelepiped(Parallelepiped):
+    length = Quantity(
+        type=float,
+        description='The y dimension of the parallelepiped. '
+        'The dimension of the sample along the path of the light beam.',
+        a_eln=dict(
+            component='NumberEditQuantity',
+            defaultDisplayUnit='millimeter',
+            label='Length (y)',
+        ),
+        unit='meter',
+    )
+
+
+class PhysicalProperties(ArchiveSection):
+    """
+    Defines the physical properties of the sample used in the transmission measurement.
+    Can be either a solid sample or liquid sample.
+    """
+
+
+class SolidSamplePhysicalProperties(PhysicalProperties):
+    """
+    Physical properties of the solid sample used in the transmission measurement.
+    """
+
+    geometry = SubSection(
+        section_def=TransmissionParallelepiped,
+    )
+    orientation = Quantity(
+        type=str,
+        description="""Crystallographic orientation of the sample in the direction
+        perpendicular to the light beam. Defined in terms of miller indices.
+        """,
+        a_eln={'component': 'StringEditQuantity'},
+    )
+
+
+class LiquidSamplePhysicalProperties(PhysicalProperties):
+    """
+    Physical properties of the liquid sample used in the transmission measurement.
+    Information related to the cuvette used for the liquid sample.
+    """
 
 
 class TransmissionSpectrophotometer(Instrument, EntryData):
